+++ code/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+\### Added
+
+- [#847](https://github.com/InditexTech/weavejs/issues/847) Measure tool & action
+
 ### Fixed
 
 - [#853](https://github.com/InditexTech/weavejs/issues/853) Fix issue with resize plugin when no upscaling is defined not changing stage size